Add flask test
